case_FastApi/requirements.txt
pytest
requests
pyyaml 